Added duplicate handling.
